chore: abstract base models
